Corrected accidentally changed indentation. TestDir and Extractor classes marked with __test__ variable not to collect by pytest. Moved pytest related constants to requirements.py. Added pytest and pytest-xdist to requirements.txt. Corrected DUMMY_DATASET_DIR in test_voc_format.py.
