--- tests/pytest_marking_constants/datumaro_components.py
@@ -1,6 +1,0 @@
-# Copyright (C) 2021 Intel Corporation
-#
-# SPDX-License-Identifier: MIT
-
-class DatumaroComponent:
-    Datumaro = "datumaro"
