--- tests/pytest_marking_constants/requirements.py
@@ -1,9 +1,0 @@
-# Copyright (C) 2021 Intel Corporation
-#
-# SPDX-License-Identifier: MIT
-
-
-class Requirements:
-
-    DATUM_GENERAL_REQ = "Datumaro general requirement"
-    DATUM_244 = "DATUM-244 Add Snyk integration"
