--- tests/pytest_marking_constants/skip_messages.py
@@ -1,6 +1,0 @@
-# Copyright (C) 2021 Intel Corporation
-#
-# SPDX-License-Identifier: MIT
-
-class SkipMessages:
-    NOT_IMPLEMENTED = "NOT IMPLEMENTED"
