+++ tests/requirements.py
@@ -1,0 +1,13 @@
+# Copyright (C) 2021 Intel Corporation
+#
+# SPDX-License-Identifier: MIT
+
+class DatumaroComponent:
+    Datumaro = "datumaro"
+
+class Requirements:
+    DATUM_GENERAL_REQ = "Datumaro general requirement"
+    DATUM_244 = "DATUM-244 Add Snyk integration"
+
+class SkipMessages:
+    NOT_IMPLEMENTED = "NOT IMPLEMENTED"
