+++ tests/test_RISE.py
@@ -8,8 +8,8 @@ from datumaro.components.launcher import Launcher
 from datumaro.components.algorithms.rise import RISE
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 
 @pytest.mark.components(DatumaroComponent.Datumaro)
