+++ tests/test_cifar_format.py
@@ -7,11 +7,11 @@ from datumaro.components.extractor import (AnnotationType, DatasetItem, Label,
     LabelCategories)
 from datumaro.plugins.cifar_format import CifarConverter, CifarImporter
 from datumaro.util.image import Image
-from datumaro.util.test_utils import TempTestDir, compare_datasets
+from datumaro.util.test_utils import TestDir, compare_datasets
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 
 @pytest.mark.components(DatumaroComponent.Datumaro)
@@ -34,7 +34,7 @@ class CifarFormatTest(TestCase):
             )
         ], categories=['label_0', 'label_1'])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             CifarConverter.convert(source_dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'cifar')
 
@@ -58,7 +58,7 @@ class CifarFormatTest(TestCase):
                 'label' + str(label) for label in range(2)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             CifarConverter.convert(source_dataset, test_dir, save_images=False)
             parsed_dataset = Dataset.import_from(test_dir, 'cifar')
 
@@ -83,7 +83,7 @@ class CifarFormatTest(TestCase):
                 'label' + str(label) for label in range(2)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             CifarConverter.convert(source_dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'cifar')
 
@@ -101,7 +101,7 @@ class CifarFormatTest(TestCase):
             ),
         ], categories=['label_0'])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             CifarConverter.convert(source_dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'cifar')
 
@@ -119,7 +119,7 @@ class CifarFormatTest(TestCase):
                 data=np.zeros((32, 32, 3)))),
         ], categories=[])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             CifarConverter.convert(dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'cifar')
 
@@ -135,7 +135,7 @@ class CifarFormatTest(TestCase):
             DatasetItem(id='b')
         ], categories=['label_0'])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             CifarConverter.convert(dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'cifar')
