+++ tests/test_imagenet_format.py
@@ -9,11 +9,11 @@ from datumaro.components.extractor import (DatasetItem, Label,
 )
 from datumaro.plugins.imagenet_format import ImagenetConverter, ImagenetImporter
 from datumaro.util.image import Image
-from datumaro.util.test_utils import TempTestDir, compare_datasets
+from datumaro.util.test_utils import TestDir, compare_datasets
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 @pytest.mark.components(DatumaroComponent.Datumaro)
 class ImagenetFormatTest(TestCase):
@@ -35,7 +35,7 @@ class ImagenetFormatTest(TestCase):
                 'label_' + str(label) for label in range(2)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             ImagenetConverter.convert(source_dataset, test_dir, save_images=True)
 
             parsed_dataset = Dataset.import_from(test_dir, 'imagenet')
@@ -60,7 +60,7 @@ class ImagenetFormatTest(TestCase):
                 'label_' + str(label) for label in range(3)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             ImagenetConverter.convert(source_dataset, test_dir, save_images=True)
 
             parsed_dataset = Dataset.import_from(test_dir, 'imagenet')
@@ -82,7 +82,7 @@ class ImagenetFormatTest(TestCase):
                 'label_' + str(label) for label in range(2)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             ImagenetConverter.convert(source_dataset, test_dir, save_images=True)
 
             parsed_dataset = Dataset.import_from(test_dir, 'imagenet')
@@ -101,7 +101,7 @@ class ImagenetFormatTest(TestCase):
                 data=np.zeros((3, 4, 3)))),
         ], categories=[])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             ImagenetConverter.convert(dataset, test_dir, save_images=True)
 
             parsed_dataset = Dataset.import_from(test_dir, 'imagenet')
