+++ tests/test_masks.py
@@ -6,8 +6,8 @@ import datumaro.util.mask_tools as mask_tools
 from datumaro.components.extractor import CompiledMask
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 
 @pytest.mark.components(DatumaroComponent.Datumaro)
