+++ tests/test_mnist_format.py
@@ -7,11 +7,11 @@ from datumaro.components.extractor import (AnnotationType, DatasetItem, Label,
     LabelCategories)
 from datumaro.plugins.mnist_format import MnistConverter, MnistImporter
 from datumaro.util.image import Image
-from datumaro.util.test_utils import TempTestDir, compare_datasets
+from datumaro.util.test_utils import TestDir, compare_datasets
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 @pytest.mark.components(DatumaroComponent.Datumaro)
 class MnistFormatTest(TestCase):
@@ -36,7 +36,7 @@ class MnistFormatTest(TestCase):
                 str(label) for label in range(10)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             MnistConverter.convert(source_dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'mnist')
 
@@ -60,7 +60,7 @@ class MnistFormatTest(TestCase):
                 str(label) for label in range(10)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             MnistConverter.convert(source_dataset, test_dir, save_images=False)
             parsed_dataset = Dataset.import_from(test_dir, 'mnist')
 
@@ -83,7 +83,7 @@ class MnistFormatTest(TestCase):
                 str(label) for label in range(10)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             MnistConverter.convert(source_dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'mnist')
 
@@ -104,7 +104,7 @@ class MnistFormatTest(TestCase):
                 str(label) for label in range(10)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             MnistConverter.convert(source_dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'mnist')
 
@@ -125,7 +125,7 @@ class MnistFormatTest(TestCase):
                 str(label) for label in range(10)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             MnistConverter.convert(dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'mnist')
 
@@ -144,7 +144,7 @@ class MnistFormatTest(TestCase):
                 str(label) for label in range(10)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             MnistConverter.convert(dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'mnist')
 
@@ -165,7 +165,7 @@ class MnistFormatTest(TestCase):
                 'label_%s' % label for label in range(2)),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             MnistConverter.convert(dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'mnist')
