+++ tests/test_mots_format.py
@@ -8,12 +8,12 @@ from datumaro.components.extractor import DatasetItem, Mask
 from datumaro.components.dataset import Dataset
 from datumaro.plugins.mots_format import MotsPngConverter, MotsImporter
 from datumaro.util.image import Image
-from datumaro.util.test_utils import (TempTestDir, compare_datasets,
-                                      check_save_and_load)
+from datumaro.util.test_utils import (TestDir, compare_datasets,
+    check_save_and_load)
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 DUMMY_DATASET_DIR = osp.join(osp.dirname(__file__), 'assets', 'mots_dataset')
 
@@ -23,8 +23,8 @@ class MotsPngConverterTest(TestCase):
     def _test_save_and_load(self, source_dataset, converter, test_dir,
             target_dataset=None, importer_args=None, **kwargs):
         return check_save_and_load(self, source_dataset, converter, test_dir,
-                                       importer='mots',
-                                       target_dataset=target_dataset, importer_args=importer_args, **kwargs)
+            importer='mots',
+            target_dataset=target_dataset, importer_args=importer_args, **kwargs)
 
     @pytest.mark.priority_medium
     @pytest.mark.reqids(Requirements.DATUM_GENERAL_REQ)
@@ -70,7 +70,7 @@ class MotsPngConverterTest(TestCase):
             ]),
         ], categories=['a', 'b', 'c', 'd'])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             self._test_save_and_load(source,
                 partial(MotsPngConverter.convert, save_images=True),
                 test_dir, target_dataset=target)
@@ -87,7 +87,7 @@ class MotsPngConverterTest(TestCase):
             ]),
         ], categories=['a'])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             self._test_save_and_load(source,
                 partial(MotsPngConverter.convert, save_images=True),
                 test_dir, require_images=True)
@@ -113,7 +113,7 @@ class MotsPngConverterTest(TestCase):
             ),
         ], categories=['a'])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             self._test_save_and_load(expected,
                 partial(MotsPngConverter.convert, save_images=True),
                 test_dir, require_images=True)
