+++ tests/test_ndr.py
@@ -9,8 +9,8 @@ from datumaro.components.extractor import (DatasetItem, Label,
 import datumaro.plugins.ndr as ndr
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 
 @pytest.mark.components(DatumaroComponent.Datumaro)
