+++ tests/test_splitter.py
@@ -17,8 +17,8 @@ import datumaro.plugins.splitter as splitter
 from datumaro.components.operations import compute_ann_statistics
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 
 @pytest.mark.components(DatumaroComponent.Datumaro)
