+++ tests/test_transforms.py
@@ -12,8 +12,8 @@ import datumaro.plugins.transforms as transforms
 from datumaro.util.test_utils import compare_datasets
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 
 @pytest.mark.components(DatumaroComponent.Datumaro)
