+++ tests/test_validator.py
@@ -21,8 +21,8 @@ from datumaro.components.validator import (ClassificationValidator,
     SegmentationValidator)
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 
 class TestValidatorTemplate(TestCase):
