+++ tests/test_vgg_face2_format.py
@@ -8,11 +8,11 @@ from datumaro.components.extractor import (AnnotationType, Bbox, DatasetItem,
 from datumaro.plugins.vgg_face2_format import (VggFace2Converter,
     VggFace2Importer)
 from datumaro.util.image import Image
-from datumaro.util.test_utils import TempTestDir, compare_datasets
+from datumaro.util.test_utils import TestDir, compare_datasets
 
 import pytest
-from tests.pytest_marking_constants.requirements import Requirements
-from tests.pytest_marking_constants.datumaro_components import DatumaroComponent
+from tests.requirements import Requirements
+from tests.requirements import DatumaroComponent
 
 
 @pytest.mark.components(DatumaroComponent.Datumaro)
@@ -57,7 +57,7 @@ class VggFace2FormatTest(TestCase):
                 [('label_%s' % i, 'class_%s' % i) for i in range(5)]),
         })
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             VggFace2Converter.convert(source_dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'vgg_face2')
 
@@ -77,7 +77,7 @@ class VggFace2FormatTest(TestCase):
             ),
         ], categories=['a'])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             VggFace2Converter.convert(source_dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'vgg_face2')
 
@@ -96,7 +96,7 @@ class VggFace2FormatTest(TestCase):
             ),
         ], categories=['a'])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             VggFace2Converter.convert(source_dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'vgg_face2')
 
@@ -117,7 +117,7 @@ class VggFace2FormatTest(TestCase):
             ),
         ], categories=['label_0'])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             VggFace2Converter.convert(source_dataset, test_dir, save_images=False)
             parsed_dataset = Dataset.import_from(test_dir, 'vgg_face2')
 
@@ -142,7 +142,7 @@ class VggFace2FormatTest(TestCase):
             ),
         ], categories=[])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             VggFace2Converter.convert(source_dataset, test_dir, save_images=False)
             parsed_dataset = Dataset.import_from(test_dir, 'vgg_face2')
 
@@ -166,7 +166,7 @@ class VggFace2FormatTest(TestCase):
             ),
         ], categories=[])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             VggFace2Converter.convert(source_dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'vgg_face2')
 
@@ -189,7 +189,7 @@ class VggFace2FormatTest(TestCase):
             ),
         ], categories=['a'])
 
-        with TempTestDir() as test_dir:
+        with TestDir() as test_dir:
             VggFace2Converter.convert(dataset, test_dir, save_images=True)
             parsed_dataset = Dataset.import_from(test_dir, 'vgg_face2')
